feat: implement suggestion relevance mapping and URL normalization utilities for search providers
