+++ src/renderer/src/lib/omnibox-new/search-providers/suggestion-utils.ts
@@ -1,0 +1,3 @@
+export function mapSuggestionRelevanceByIndex(index: number): number {
+  return Math.max(100, 400 - index * 40);
+}
